Move FlixelBar to utility package
